Finish moving files and functions around
